requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
asyncpg==0.30.0
alembic==1.16.4
pydantic-settings==2.10.1
PyJWT==2.10.1
bcrypt==4.3.0
email-validator==2.2.0
python-multipart==0.0.20
